src/backend/huggingface/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
-f https://download.pytorch.org/whl/torch_stable.html
semantic-kernel==0.3.13.dev0
torch==2.1.0+cu121
transformers
nltk
azure-identity==1.12.0
openai==0.28.1
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
